* add python files for submitting CRAB jobs * add data files from Run2_2016 * remove unused variable "mother" from GetGenHSCPBeta (to allow for compilation by scram) * pass base directory by string to InitBaseDirectory() function * clean Analysis_Samples.txt. This fie is now created automatically inside a CRAB job * fix int <-> usigned int type mismatch in some places * change name of Analysis_Step1_EventLoop(char* SavePath) to EventLoop(char* SavePath) * pass base directory by mode type (previously it was always "ANALYSIS") * initialize edm::LumiReWeighting only for MC jobs, as it crashes for data at the end of job (when LumiReWeighting is destroyed) * rename output hitos to sinmply "Histos.root" to allows easier retrieval * add a hack for missig det units in getOuterHitPos(). The CMS_GeomTree.root has to be updated * add some small fixes to follow UI change for TrigerResult retrieval in CMSSW * comment two broken lines with   //if(DXYSB && DZSB)st->BS_Cosmic->Fill(OpenAngle,Event_Weight);//TEST AK   //if(TypeMode==5 && fabs(dz)>GlobalMaxDZ) DZSB = true;//TEST AK * remove "/../../" from paths to analysis input ROOT files, as this does not work on a remote site. * split lines with merged "if" statements (to be verified by HSCP expert)
if(HSCPGenBeta1>=0)SamplePlots->Beta_Gen      ->Fill(HSCPGenBeta1, Event_Weight);  if(HSCPGenBeta2>=0)SamplePlots->Beta_Gen       ->Fill(HSCPGenBeta2, Event_Weight);
if(HSCPGenBeta1>=0)SamplePlots->Beta_GenCharged->Fill(HSCPGenBeta1, Event_Weight); if(HSCPGenBeta2>=0)SamplePlots->Beta_GenCharged->Fill(HSCPGenBeta2, Event_Weight);

+++ test/AnalysisCode/Analysis_TOFUtility.h
@@ -693,7 +693,7 @@ class muonTimingCalculator{
                     }
 
                     unsigned int segsize = seg.size();
-                    if (segsize<theHitsMin_) continue;
+                    if (segsize<(unsigned int)theHitsMin_) continue;
 
                     double a=0, b=0;
                     std::vector <double> hitxl,hitxr,hityl,hityr;
